webapi: specify authorization attribute run order
Roslyn emits attributes in a different order than the old compiler,
so build ordering in ourselves to make sure it's consistent across
versions.

There is probably a better way to do this, but just solving it with
code instead of other methods via configuration/wire-up which may
not be backwards compatible.